merge google 2 local

+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mp3/VbriSeeker.java
@@ -89,7 +89,7 @@ import com.google.android.exoplayer2.util.Util;
         default:
           return null;
       }
-      position += segmentSize * scale;
+      position += segmentSize * ((long) scale);
     }
     if (inputLength != C.LENGTH_UNSET && inputLength != position) {
       Log.w(TAG, "VBRI data size mismatch: " + inputLength + ", " + position);
